streamlit_app/requirements.txt
streamlit==1.37.1
python-frontmatter==1.1.0
PyYAML==6.0.2
